Some code clean up in ProgramTesr
