+++ ProgramTest/ProgramTest/CreatePreset.cs
@@ -1,16 +1,7 @@
 ﻿using Data.Entities;
 using Data.Repositories;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-using System.Data.Entity;
-
 
 namespace ProgramTest
 {
@@ -21,10 +12,6 @@ namespace ProgramTest
         {
             _presetRepository = new PresetRepository(Program.DbContext);
             InitializeComponent();
-        }
-        private void label1_Click(object sender, EventArgs e)
-        {
-
         }
 
         private void button2_Click(object sender, EventArgs e)
